Refactor Timer: modularización con respuestas estandarizadas  agreger un timerController.js
